requirements.txt
spacy
pyenchant
pandas
numpy
scikit-learn